Implement counting of lines and of whitespace characters per line.
The characters in comments are counted as whitespace.
This is the first part of the work to decorate the ast
with original location information.

+++ src/parse/parser.rs
@@ -6,8 +6,8 @@ use parse::ast;
 //use parse::lexer;
 
 pub struct Parser {
-    pub line: i64,
-    pub column: i64,
+    pub line: usize,
+    pub column: usize,
 }
 impl Parser {
